atualiza licao 14

+++ bd-projeto-guiado-III/gameStore/src/controller/gamesController.js
@@ -1,10 +1,24 @@
 const GamesModel = require("../models/gamesModel");
 const ConsolesModel = require("../models/consolesModel");
+const jwt = require("jsonwebtoken");
+const SECRET = process.eventNames.SECRET;
 
 const findAllGames = async (req, res) => {
   try {
-    const allGames = await GamesModel.find().populate("console");
-    res.status(200).json(allGames);
+    const authHeader = req.get("authorization")
+    if (!authHeader) {
+      return res.status(401).send("informações de autorização não passada")
+    }
+    const token = authHeader.split("")[1]
+
+    jwt.verify(token, SECRET, async function (erro) {
+      if (erro) {
+        return res.status(403).send("acesso não autorizado")
+      }
+      const allGames = await GamesModel.find().populate("console");
+      res.status(200).json(allGames);
+    })
+
   } catch {
     res.status(500).json({ message: error.message });
   };
@@ -12,6 +26,19 @@ const findAllGames = async (req, res) => {
 
 const findGameById = async (req, res) => {
   try {
+    const authHeader = req.get("authorization")
+    if (!authHeader) {
+      return res.status(401).send("Não tem nada aqui")
+    }
+
+    const token = authHeader(" ")[1]
+
+    jwt.verify(token, SECRET, async function (erro){
+      if (erro){
+        return res.status(403).send("Sem permissão de acesso")
+  }
+  
+})
     const findGame = await GamesModel.findById(req.params.id).populate(
       "console"
     );
@@ -20,49 +47,63 @@ const findGameById = async (req, res) => {
     }
     res.status(200).json(findGame);
   } catch (error) {
+
     res.status(500).json({ message: error.message });
   };
 };
 
 const addNewGame = async (req, res) => {
   try {
-    const {
-      consoleId,
-      name,
-      developer,
-      releaseDate,
-      genre,
-      mode,
-      available,
-      description,
-    } = req.body;
+    const authHeader = req.get("authorization")
+    if (!authHeader) {
+      return res.status(401).send("Nã há cabeçalho quirida")
+    }
 
-    if (!consoleId) {
-      return res
-        .status(400)
-        .json({ message: "Required: Enter the Console id." });
-    };
+    const token = authHeader - split(" ")[1]
 
-    const findConsole = await ConsolesModel.findById(consoleId);
+    jwt.verify(token, SECRET, async function (erro) {
+      if (erro) {
+        return res.status(403).send("Você não pode acessar")
+      }
+      const {
+        consoleId,
+        name,
+        developer,
+        releaseDate,
+        genre,
+        mode,
+        available,
+        description,
+      } = req.body;
 
-    if (!findConsole) {
-      return res.status(404).json({ message: "Console not found" });
-    };
+      if (!consoleId) {
+        return res
+          .status(400)
+          .json({ message: "Required: Enter the Console id." });
+      };
 
-    const newGame = new GamesModel({
-      console: consoleId,
-      name,
-      developer,
-      releaseDate,
-      genre,
-      mode,
-      available,
-      description,
-    });
-    const savedGame = await newGame.save();
-    res
-      .status(200)
-      .json({ message: "New game added successfully!", savedGame });
+      const findConsole = await ConsolesModel.findById(consoleId);
+
+      if (!findConsole) {
+        return res.status(404).json({ message: "Console not found" });
+      };
+
+      const newGame = new GamesModel({
+        console: consoleId,
+        name,
+        developer,
+        releaseDate,
+        genre,
+        mode,
+        available,
+        description,
+      });
+      const savedGame = await newGame.save();
+      res
+        .status(200)
+        .json({ message: "New game added successfully!", savedGame });
+    })
+
   } catch (error) {
     console.error(error);
     res.status(500).json({ message: error.message });
@@ -71,40 +112,53 @@ const addNewGame = async (req, res) => {
 
 const updateGame = async (req, res) => {
   try {
-    const { id } = req.params;
-    const {
-      consoleId,
-      name,
-      developer,
-      releaseDate,
-      genre,
-      mode,
-      available,
-      description,
-    } = req.body;
-    const findGame = await GamesModel.findById(id);
-    if (findGame == null) {
-      res.status(404).json({ message: "Game not found" });
-    };
+    const authHeader = req.get("authorization")
+    if (!authHeader) {
+      return res.status(401).send("Não tem autorização")
+    }
 
-    if (consoleId) {
-      const findConsole = await ConsolesModel.findById(consoleId);
+    const token = authHeader(" ")[1]
 
-      if (findConsole == null) {
-        return res.status(404).json({ message: "Console not found" });
+    jwt.verify(token, SECRET, async function (erro) {
+      if (erro) {
+        return res.status(403).send("Sem permissão de acesso")
+      }
+      const { id } = req.params;
+      const {
+        consoleId,
+        name,
+        developer,
+        releaseDate,
+        genre,
+        mode,
+        available,
+        description,
+      } = req.body;
+      const findGame = await GamesModel.findById(id);
+      if (findGame == null) {
+        res.status(404).json({ message: "Game not found" });
       };
-    };
-    findGame.name = name || findGame.name;
-    findGame.developer = developer || findGame.developer;
-    findGame.releaseDate = releaseDate || findGame.releaseDate;
-    findGame.genre = genre || findGame.genre;
-    findGame.mode = mode || findGame.mode;
-    findGame.available = available || findGame.available;
-    findGame.description = description || findGame.description;
-    findGame.console = consoleId || findGame.console;
 
-    const savedGame = await findGame.save();
-    res.status(200).json({ message: "Game successfully updated", savedGame });
+      if (consoleId) {
+        const findConsole = await ConsolesModel.findById(consoleId);
+
+        if (findConsole == null) {
+          return res.status(404).json({ message: "Console not found" });
+        };
+      };
+      findGame.name = name || findGame.name;
+      findGame.developer = developer || findGame.developer;
+      findGame.releaseDate = releaseDate || findGame.releaseDate;
+      findGame.genre = genre || findGame.genre;
+      findGame.mode = mode || findGame.mode;
+      findGame.available = available || findGame.available;
+      findGame.description = description || findGame.description;
+      findGame.console = consoleId || findGame.console;
+
+      const savedGame = await findGame.save();
+      res.status(200).json({ message: "Game successfully updated", savedGame })
+    })
+
   } catch (error) {
     res.status(500).json({ message: error.message });
   };
@@ -112,14 +166,27 @@ const updateGame = async (req, res) => {
 
 const deleteGame = async (req, res) => {
   try {
-    const { id } = req.params;
-    const findGames = await GamesModel.findByIdAndDelete(id);
+    const authHeader = req.get("authorization")
+    if (!authHeader) {
+      return res.status(401).send("Esqueceu a autorização")
+    }
 
-    if (findGames == null) {
-      return res.status(404).json({ message: `Game with id ${id} not found` })
-    };
-    
-    res.status(200).json({ message: `Game with id ${id} was successfully deleted` });
+    const token = authHeader(" ")[1]
+
+    jwt.verify(token, SECRET, async function (erro) {
+      if (erro) {
+        return res.status(403).send("Não vai rolar")
+      }
+      const { id } = req.params;
+      const findGames = await GamesModel.findByIdAndDelete(id);
+
+      if (findGames == null) {
+        return res.status(404).json({ message: `Game with id ${id} not found` })
+      };
+
+      res.status(200).json({ message: `Game with id ${id} was successfully deleted` });
+    })
+
   } catch (error) {
     res.status(500).json({ message: error.message });
   };
